refactor: move shared auth dependencies into the common package
Extract the settings-free authn/authz plumbing — bearer extraction,
auth-config loading, subject-context/entitlement extraction, the
require_authn flow, and authz enforcement with the structured error
mapping — into common. Each agent keeps a thin dependencies shim that
wires its own settings, singletons, and hierarchy extractors, so route
signatures and dependency names are unchanged.

finops adopts the majority entitlement semantics (an empty claim value
yields an empty entitlement list rather than falling back to the sub
claim) and the structured error bodies for authn failures.
