+++ agents/common/pyproject.toml
@@ -4,8 +4,12 @@ version = "0.1.0"
 description = "Shared infrastructure for the OpenChoreo Python agent services"
 requires-python = ">=3.14"
 dependencies = [
+    "fastapi>=0.139.2",
+    "httpx>=0.23.0",
     "jinja2>=3.1.0",
     "pydantic>=2.0.0",
+    "pyjwt[crypto]>=2.13.0",
+    "pyyaml>=6.0.0",
 ]
 
 [tool.ruff]
